tests/wpt/web-platform-tests/dom/nodes/aria-element-reflection.tentative.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Element Reflection for aria-activedescendant and aria-errormessage</title>
    <link rel=help href="https://whatpr.org/html/3917/common-dom-interfaces.html#reflecting-content-attributes-in-idl-attributes:element">
    <link rel="author" title="Meredith Lane" href="meredithl@chromium.org">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <div id="activedescendant" aria-activedescendant="x"></div>

  <div id="parent-listbox" role="listbox" aria-activedescendant="i1">
    <div role="option" id="i1">Item 1</div>
    <div role="option" id="i2">Item 2</div>
  </div>

  <script>
  test(function(t) {
    const ancestor = document.getElementById("parent-listbox");
    assert_equals(activedescendant.ariaActiveDescendantElement, null,
                  "invalid ID for relationship returns null");

    const descendant1 = document.getElementById("i1");
    const descendant2 = document.getElementById("i2");

    // Element reference should be set if the content attribute was included.
    assert_equals(ancestor.getAttribute("aria-activedescendant"), "i1", "check content attribute after parsing.");
    assert_equals(ancestor.ariaActiveDescendantElement, i1, "check idl attribute after parsing.");

    // If we set the content attribute, the element reference should reflect this.
    ancestor.setAttribute("aria-activedescendant", "i2");
    assert_equals(ancestor.ariaActiveDescendantElement, descendant2, "setting the content attribute updates the element reference.");

    // Setting the element reference should be reflected in the content attribute.
    ancestor.ariaActiveDescendantElement = descendant1;
    assert_equals(ancestor.ariaActiveDescendantElement, descendant1, "getter should return the right element reference.");
    assert_equals(ancestor.getAttribute("aria-activedescendant"), "i1", "content attribute should reflect the element reference.");

    // Both content and IDL attribute should be nullable.
    ancestor.ariaActiveDescendantElement = null;
    assert_equals(ancestor.ariaActiveDescendantElement, null);
    assert_false(ancestor.hasAttribute("aria-activedescendant"));
    assert_equals(ancestor.getAttribute("aria-activedescendant"), null, "nullifying the idl attribute removes the content attribute.");

    // Setting content attribute to non-existent or non compatible element should nullify the IDL attribute.
    // Reset the element to an existant one.
    ancestor.setAttribute("aria-activedescendant", "i1");
    assert_equals(ancestor.ariaActiveDescendantElement, i1, "reset attribute.");

    ancestor.setAttribute("aria-activedescendant", "non-existent-element");
    assert_equals(ancestor.getAttribute("aria-activedescendant"), "non-existent-element");
    assert_equals(ancestor.ariaActiveDescendantElement, null,"non-DOM content attribute should null the element reference");
  }, "aria-activedescendant element reflection");
  </script>

  <div id="parent-listbox-2" role="listbox" aria-activedescendant="option1">
    <div role="option" id="option1">Item 1</div>
    <div role="option" id="option2">Item 2</div>
  </div>

  <script>
  test(function(t) {
    const ancestor = document.getElementById("parent-listbox-2");
    const option1 = document.getElementById("option1");
    const option2 = document.getElementById("option2");
    assert_equals(ancestor.ariaActiveDescendantElement, option1);

    option1.removeAttribute("id");
    option2.setAttribute("id", "option1");
    const option2Duplicate = document.getElementById("option1");
    assert_equals(option2, option2Duplicate);

    assert_equals(ancestor.ariaActiveDescendantElement, option2);
  }, "If the content attribute is set directly, the IDL attribute getter always returns the first element whose ID matches the content attribute.");
  </script>

  <div id="blank-id-parent" role="listbox">
    <div role="option" id="multiple-id"></div>
    <div role="option" id="multiple-id"></div>
  </div>

  <script>
  test(function(t) {
    const ancestor = document.getElementById("blank-id-parent");

    // Get second child of parent. This violates the setting of a reflected element
    // as it will not be the first child of the parent with that ID, which should
    // result in an empty string for the content attribute.
    ancestor.ariaActiveDescendantElement = ancestor.children[1];
    assert_true(ancestor.hasAttribute("aria-activedescendant"));
    assert_equals(ancestor.getAttribute("aria-activedescendant"), "");
    assert_equals(ancestor.ariaActiveDescendantElement, ancestor.children[1]);
  }, "Setting the IDL attribute to an element which is not the first element in DOM order with its ID causes the content attribute to be an empty string");
  </script>

  <div id="outer-container">
    <p id="light-paragraph">Hello world!</p>
    <span class="shadow-host">
    </span>
  </div>

  <script>
  test(function(t) {
    const shadowElement = document.querySelector(".shadow-host");
    const shadow = shadowElement.attachShadow({mode: "open"});
    const link = document.createElement("a");
    shadow.appendChild(link);

    const lightPara = document.getElementById("light-paragraph");
    assert_equals(lightPara.ariaActiveDescendantElement, null);

    // The given element crosses a shadow dom boundary, so it cannot be
    // set as an element reference.
    lightPara.ariaActiveDescendantElement = link;
    assert_equals(lightPara.ariaActiveDescendantElement, null);

    // The given element crosses a shadow dom boundary (upwards), so
    // can be used as an element reference, but the content attribute
    // should reflect the empty string.
    link.ariaActiveDescendantElement = lightPara;
    assert_equals(link.ariaActiveDescendantElement, lightPara);
    assert_equals(link.getAttribute("aria-activedescendant"), "");
  }, "Setting an element reference that crosses into a shadow tree is disallowed, but setting one that is in a shadow inclusive ancestor is allowed.");
  </script>

  <input id="startTime" ></input>
  <span id="errorMessage">Invalid Time</span>

  <script>
  test(function(t) {
    const inputElement = document.getElementById("startTime");
    const errorMessage = document.getElementById("errorMessage");

    inputElement.ariaErrorMessageElement = errorMessage;
    assert_equals(inputElement.getAttribute("aria-errormessage"), "errorMessage");

    inputElement.ariaErrorMessageElement = null;
    assert_equals(inputElement.ariaErrorMessageElement, null, "blah");
    assert_false(inputElement.hasAttribute("aria-errormessage"));

    inputElement.setAttribute("aria-errormessage", "errorMessage");
    assert_equals(inputElement.ariaErrorMessageElement, errorMessage);

  }, "aria-errormessage");

  </script>

  <label>
    Password:
    <input id="password-field" type="password" aria-details="pw">
  </label>

  <ul>
    <li id="list-item-1">First description.</li>
    <li id="list-item-2">Second description.</li>
  </ul>

  <script>

  test(function(t) {
    const inputElement = document.getElementById("password-field");
    const ul1 = document.getElementById("list-item-1");
    const ul2 = document.getElementById("list-item-2");

    assert_equals(inputElement.ariaDetailsElement, null);
    inputElement.ariaDetailsElement = ul1;
    assert_equals(inputElement.getAttribute("aria-details"), "list-item-1");

    inputElement.ariaDetailsElement = ul2;
    assert_equals(inputElement.getAttribute("aria-details"), "list-item-2");
  }, "aria-details");
  </script>

  <div id="old-parent" role="listbox" aria-activedescendant="content-attr-element">
    <div role="option" id="content-attr-element">Item 1</div>
    <div role="option" id="idl-attr-element">Item 2</div>
  </div>

  <script>

  test(function(t) {
    const deletionParent = document.getElementById("old-parent");

    const descendant1 = document.getElementById("content-attr-element");
    const descendant2 = document.getElementById("idl-attr-element");

    // Deleting an element set via the content attribute.
    assert_equals(deletionParent.getAttribute("aria-activedescendant"), "content-attr-element");
    assert_equals(deletionParent.ariaActiveDescendantElement, descendant1);

    deletionParent.removeChild(descendant1);
    assert_equals(deletionParent.getAttribute("aria-activedescendant"), "content-attr-element");
    assert_equals(deletionParent.ariaActiveDescendantElement, null);

    // Deleting an element set via the IDL attribute.
    deletionParent.ariaActiveDescendantElement = descendant2;
    assert_equals(deletionParent.getAttribute("aria-activedescendant"), "idl-attr-element");

    deletionParent.removeChild(descendant2);
    assert_equals(deletionParent.ariaActiveDescendantElement, null);

    // The content attribute will still reflect the id.
    assert_equals(deletionParent.getAttribute("aria-activedescendant"), "idl-attr-element");
  }, "Deleting a reflected element should return null for the IDL attribute and cause the content attribute to become stale.");
  </script>

  <div id="parent" role="listbox" aria-activedescendant="original">
    <div role="option" id="original">Item 1</div>
    <div role="option" id="element-with-persistant-id">Item 2</div>
  </div>

  <script>
  test(function(t) {
    const parentNode = document.getElementById("parent");
    const changingIdElement = document.getElementById("original");
    const persistantIDElement = document.getElementById("element-with-persistant-id");

    assert_equals(parentNode.ariaActiveDescendantElement, changingIdElement);

    // Modify the id attribute.
    changingIdElement.setAttribute("id", "new-id");

    // The content attribute still reflects the old id, and we expect the
    // Element reference to be null as there is no DOM node with id "original"
    assert_equals(parentNode.getAttribute("aria-activedescendant"), "original");
    assert_equals(parentNode.ariaActiveDescendantElement, null, "Element set via content attribute with a changed id will return null on getting");

    parentNode.ariaActiveDescendantElement = changingIdElement;
    assert_equals(parentNode.getAttribute("aria-activedescendant"), "new-id");

    // The explicitly set element takes precendance over the content attribute.
    // This means that we still return the same element reference, but the
    // content attribute reflects the old id.
    changingIdElement.setAttribute("id", "newer-id");
    assert_equals(parentNode.ariaActiveDescendantElement, changingIdElement, "explicitly set element is still present even after the id has been changed");
    assert_equals(parentNode.getAttribute("aria-activedescendant"), "new-id", "content attribute reflects the id that was present upon explicitly setting the element reference.");
  }, "Changing the ID of an element causes the content attribute to become out of sync.");
  </script>

  <div id="light-parent" role="listbox">
    <div role="option" id="light-element">Hello world!</div>
  </div>
  <div id="shadowHost"></div>

  <script>
  test(function(t) {
    const shadowElement = document.getElementById("shadowHost");
    const shadowHost = shadowElement.attachShadow({mode: "open"});
    const lightParent = document.getElementById("light-parent");
    const optionElement = document.getElementById("light-element");

    lightParent.ariaActiveDescendantElement = optionElement;
    assert_equals(lightParent.ariaActiveDescendantElement, optionElement);

    // Move the referenced element into shadow DOM.
    shadowHost.appendChild(optionElement);
    assert_equals(lightParent.ariaActiveDescendantElement, null);
    assert_equals(lightParent.getAttribute("aria-activedescendant"), "light-element");

    // Move the referenced element back into light DOM.
    lightParent.appendChild(optionElement);
    assert_equals(lightParent.ariaActiveDescendantElement, optionElement);
  }, "Reparenting an element into a descendant shadow scope nullifies the element reference.");
  </script>

  <div id="myBillingId">Billing</div>

  <div>
      <div id="myNameId">Name</div>
      <input type="text" id="input1" aria-labelledby="myBillingId myNameId"/>
  </div>
  <div>
      <div id="myAddressId">Address</div>
      <input type="text" id="input2"/>
  </div>

  <script>
  test(function(t) {
    const billingElement = document.getElementById("myBillingId");
    const nameElement = document.getElementById("myNameId");
    const addressElement = document.getElementById("myAddressId");
    const input1 = document.getElementById("input1");
    const input2 = document.getElementById("input2");

    assert_array_equals(input1.ariaLabelledByElements, [billingElement, nameElement], "parsed content attribute sets element references.");
    assert_equals(input2.ariaLabelledByElements, null, "Testing empty content attribute after parsing.");

    input2.ariaLabelledByElements = [billingElement, addressElement];
    assert_array_equals(input2.ariaLabelledByElements, [billingElement, addressElement], "Testing IDL setter/getter.");
    assert_equals(input2.getAttribute("aria-labelledby"), "myBillingId myAddressId");

    billingElement.remove();
    assert_array_equals(input2.ariaLabelledByElements, [addressElement]);

    input2.ariaLabelledByElements = [];
    assert_array_equals(input2.ariaLabelledByElements, [], "Testing IDL setter/getter for empty array.");
    assert_equals(input2.getAttribute("aria-labelledby"), "");

    input1.removeAttribute("aria-labelledby");
    assert_equals(input1.ariaLabelledByElements, null);
  }, "aria-labelledby.");
  </script>

  <ul role="tablist">
    <li role="presentation"><a id="link1" role="tab" aria-controls="panel1">Tab 1</a></li>
    <li role="presentation"><a id="link2" role="tab">Tab 2</a></li>
  </ul>

  <div role="tabpanel" id="panel1"></div>
  <div role="tabpanel" id="panel2"></div>

  <script>
  test(function(t) {
    const link1 = document.getElementById("link1");
    const link2 = document.getElementById("link2");
    const panel1 = document.getElementById("panel1");
    const panel2 = document.getElementById("panel2");
    assert_array_equals(link1.ariaControlsElements, [panel1]);
    assert_equals(link2.ariaControlsElements, null);

    link2.setAttribute("aria-controls", "panel1 panel2");
    assert_array_equals(link2.ariaControlsElements, [panel1, panel2]);

    link1.ariaControlsElements = [];
    assert_equals(link1.getAttribute("aria-controls"), "");

    link2.ariaControlsElements = [panel1, panel2];
    assert_equals(link2.getAttribute("aria-controls"), "panel1 panel2");

    link2.removeAttribute("aria-controls");
    assert_equals(link2.ariaControlsElements, null);
  }, "aria-controls.");
  </script>

  <a id="described-link" aria-describedby="description1 description2">Fruit</a>
  <div id="description1">Delicious</div>
  <div id="description2">Nutritious</div>

  <script>
  test(function(t) {
    const link = document.getElementById("described-link");
    const description1 = document.getElementById("description1");
    const description2 = document.getElementById("description2");
    assert_array_equals(link.ariaDescribedByElements, [description1, description2]);

    link.ariaDescribedByElements = [description1, description2];
    assert_equals(link.getAttribute("aria-describedby"), "description1 description2");

    link.ariaDescribedByElements = [];
    assert_equals(link.getAttribute("aria-describedby"), "");

    link.setAttribute("aria-describedby", "description1");
    assert_array_equals(link.ariaDescribedByElements, [description1]);

    link.removeAttribute("aria-describedby");
    assert_equals(link.ariaDescribedByElements, null);
  }, "aria-describedby.");
  </script>

  <h2 id="title-heading" aria-flowto="article1 article2">Title</h2>
  <div>Next</div>
  <article id="article2">Content2</article>
  <article id="article1">Content1</article>

  <script>
  test(function(t) {
    const heading = document.getElementById("title-heading");
    const article1 = document.getElementById("article1");
    const article2 = document.getElementById("article2");

    assert_array_equals(heading.ariaFlowToElements, [article1, article2]);

    heading.ariaFlowToElements = [article1, article2];
    assert_equals(heading.getAttribute("aria-flowto"), "article1 article2");

    heading.ariaFlowToElements = [];
    assert_equals(heading.getAttribute("aria-flowto"), "");

    heading.setAttribute("aria-flowto", "article1");
    assert_array_equals(heading.ariaFlowToElements, [article1]);

    heading.removeAttribute("aria-flowto");
    assert_equals(heading.ariaFlowToElements, null);
  }, "aria-flowto.");
  </script>

  <ul>
    <li id="li-owner" aria-owns="child1 child2">Parent</li>
  </ul>
  <ul>
    <li id="child1">Child 1</li>
    <li id="child2">Child 2</li>
  </ul>
  <script>
  test(function(t) {
    const owner = document.getElementById("li-owner");
    const child1 = document.getElementById("child1");
    const child2 = document.getElementById("child2");

    assert_array_equals(owner.ariaOwnsElements, [child1, child2]);

    owner.removeAttribute("aria-owns");
    assert_equals(owner.ariaOwnsElements, null);

    owner.ariaOwnsElements = [child1, child2];
    assert_equals(owner.getAttribute("aria-owns"), "child1 child2");

    owner.ariaOwnsElements = [];
    assert_equals(owner.getAttribute("aria-owns"), "");

    owner.setAttribute("aria-owns", "child1");
    assert_array_equals(owner.ariaOwnsElements, [child1]);
  }, "aria-owns.");
  </script>

  <div id="light-dom-container">
    <h2 id="light-dom-heading" aria-flowto="shadow-child-1 shadow-child-2">Light DOM Heading</h2>
    <div id="shadow-dom-host"></div>
    <p id="light-dom-text-1">Light DOM text</p>
    <p id="light-dom-text-2">Light DOM text</p>
  </div>

  <script>
  test(function(t) {
    const shadowHost = document.getElementById("shadow-dom-host");
    const lightDomHeading = document.getElementById("light-dom-heading");
    const lightDomText1 = document.getElementById("light-dom-text-1");
    const lightDomText2 = document.getElementById("light-dom-text-2");

    const shadowRoot = shadowHost.attachShadow({mode: "open"});
    const shadowChild1 = document.createElement("article");
    shadowChild1.setAttribute("id", "shadow-child-1");
    shadowRoot.appendChild(shadowChild1);
    const shadowChild2 = document.createElement("article");
    shadowChild2.setAttribute("id", "shadow-child-1");
    shadowRoot.appendChild(shadowChild2);

    // The elements in the content attribute are in a "darker" tree - they
    // enter a shadow encapsulation boundary, so should reflect null for
    // the content attribute.
    assert_equals(lightDomHeading.ariaFlowToElements, null);

    // These elements are in a shadow including ancestor, i.e "lighter" tree.
    // Valid for the IDL attribute, but content attribute should be null.
    shadowChild1.ariaFlowToElements = [lightDomText1, lightDomText2];
    assert_equals(shadowChild1.getAttribute("aria-flowto"), "", "empty content attribute for elements that cross shadow boundaries.");

    // These IDs belong to a different scope, so the attr-associated-element
    // cannot be computed.
    shadowChild2.setAttribute("aria-flowto", "light-dom-text-1 light-dom-text-2");
    assert_equals(shadowChild2.ariaFlowToElements, null);

    // Elements that cross into shadow DOM are dropped, only reflect the valid
    // elements in IDL and in the content attribute.
    lightDomHeading.ariaFlowToElements = [shadowChild1, shadowChild2, lightDomText1, lightDomText2];
    assert_array_equals(lightDomHeading.ariaFlowToElements, [lightDomText1, lightDomText2]);
    assert_equals(lightDomHeading.getAttribute("aria-flowto"), "light-dom-text-1 light-dom-text-2", "empty content attribute if any given elements cross shadow boundaries");

    // Using a mixture of elements in the same scope and in a shadow including
    // ancestor should set the IDL attribute, but should reflect the empty
    // string in the content attribute.
    shadowChild1.removeAttribute("aria-flowto");
    shadowChild1.ariaFlowToElements = [shadowChild1, lightDomText1];
    assert_equals(shadowChild1.getAttribute("aria-flowto"), "", "Setting IDL elements with a mix of scopes should reflect an empty string in the content attribute")

  }, "shadow DOM behaviour for FrozenArray element reflection.");
  </script>
</html>
